+++ frontend/src/styles/App.css
@@ -38,3 +38,7 @@ a {
 .ms-Button-flexContainer i.ms-Button-icon {
     margin-left: 14px;
 }
+
+.pd-inset-remove > * {
+    padding: 0;
+}
